backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
dj-database-url==2.3.0
whitenoise==6.9.0
gunicorn==23.0.0

drf-spectacular==0.29.0
drf-spectacular-sidecar

python-dotenv==1.2.1
psycopg2-binary==2.9.11
tzdata==2025.2

# OpenAI client (resmi kütüphane)
openai>=2.14.0

# İsteğe bağlı – sadece yerel model kullanımı için
transformers==4.57.1
torch==2.9.0
sentencepiece==0.2.1